Cleaned up the remove_com_from_hstring function
I didn't improve the algorithm of the function just made the code a bit
more rusty. The function was also returning a result, but never actually
returned any Err's. I removed the Result, and it just return the HString
not sure if that was future planning for making it better later, in
which case it probably needs to be added back in.