Improve Performance
I see no reason the Winnit event loop needs to set to Poll. It doesn't need to be sitting there running the event loop every time it finishes the loop. It was causing this app to use between 15-20% CPU on my machine. Changing it to Wait for events to occur seems to have no ill effects, but reduces the idle CPU usage for the app to 0% on my machine.